Levels - nuevo sistema

+++ Assets/sceneObjects/scripts/Bumper.cs
@@ -4,7 +4,7 @@ using System.Collections;
 public class Bumper : MonoBehaviour {
 
 	//private float _y;
-    public int force = 16;
+    public float force = 16;
     public AnimationClip animationClip;
     public bool backwardJump;
 
@@ -14,7 +14,7 @@ public class Bumper : MonoBehaviour {
 		{
 		case "Player":
 			CharacterBehavior ch = other.transform.parent.GetComponent<CharacterBehavior> ();
-			ch.SuperJumpByBumped (force * 100, 0.5f, backwardJump);
+			ch.SuperJumpByBumped ((int)force * 100, 0.5f, backwardJump);
 
 			Invoke ("Reset", 0.5f);
 			break;
